Introduction of PackageReferenceBase
