+++ src/DotNetProjectFile.Analyzers/Analyzers/MsBuild/UseAnalyzersForPackages.cs
@@ -8,13 +8,11 @@ public sealed class UseAnalyzersForPackages() : MsBuildProjectFileAnalyzer(Rule.
 
     protected override void Register(ProjectFileAnalysisContext context)
     {
-        var packageReferences = context.File.Walk().OfType<PackageReference>();
-        var globalReferences = context.File.Walk().OfType<GlobalPackageReference>();
+        var packageReferences = context.File.Walk().OfType<PackageReferenceBase>().Where(p => p is not PackageVersion);
 
         var unusedAnalyzers = Analyzers.Where(analyzer
             => analyzer.IsApplicable(context.Compilation.Options.Language)
-            && packageReferences.None(analyzer.IsMatch)
-            && globalReferences.None(analyzer.IsMatch));
+            && packageReferences.None(analyzer.IsMatch));
 
         foreach (var analyzer in unusedAnalyzers)
         {
