+++ src/DotNetProjectFile.Analyzers/MsBuild/GlobalPackageReference.cs
@@ -1,11 +1,6 @@
 namespace DotNetProjectFile.MsBuild;
 
 public sealed class GlobalPackageReference(XElement element, Node parent, MsBuildProject project)
-    : Node(element, parent, project)
+    : PackageReferenceBase(element, parent, project)
 {
-    public string? Include => Attribute();
-
-    public string? Version => Attribute();
-
-    public string? PrivateAssets => Attribute() ?? Child();
 }
