+++ src/DotNetProjectFile.Analyzers/MsBuild/PackageVersion.cs
@@ -1,13 +1,6 @@
 namespace DotNetProjectFile.MsBuild;
 
 public sealed class PackageVersion(XElement element, Node parent, MsBuildProject project)
-    : Node(element, parent, project)
+    : PackageReferenceBase(element, parent, project)
 {
-    public string? Include => Attribute();
-
-    public string? Update => Attribute();
-
-    public string? Version => Attribute();
-
-    public string IncludeOrUpdate => Include ?? Update ?? string.Empty;
 }
