+++ src/DotNetProjectFile.Analyzers/NuGet/Analyzer.cs
@@ -12,10 +12,6 @@ public sealed record Analyzer : Package
     public bool IsApplicable(string compilationLanguage)
         => Language is null || Language == compilationLanguage;
 
-    public bool IsMatch(PackageReference reference) => reference.Include.IsMatch(Name);
-
-    public bool IsMatch(GlobalPackageReference reference) => reference.Include.IsMatch(Name);
-
     public bool IsMatch(AssemblyIdentity assembly)
         => assembly.Name.StartsWith(Match, StringComparison.OrdinalIgnoreCase)
         && (assembly.Name.Length == Match.Length
